Forbid PushEdges from terminal components
When scanning a string, PushEdges don't really make a lot of sense.
PushEdges imply a return to a LabelEdge eventually (perhaps through a
chain of use_value). That is not possible since we already forbid
LableEdges and StringEdges from sharing a component.

This change will help us relax restrictions on shift-shift conflicts
through PushEdges.