Added glrParseUnique, changed glrParse return type
Two changes rolled into one because I was lazy.

Most use cases involve unambiguous grammar, so we added glrParseUnique.
This cleans up dfa_test.cpp a bit. Then we ended up changing the return
type of glrParse to include a SharedDiagSet. The returned diagnostics
are still incomplete: we should have better handling for unrecoverable
errors.

And we need to add tests for the diags we do return. But at least we
didn't break any existing stuff.