Added Diag for unexpected input char
Turns out we need to break out of parsing early in order for this to
work properly (and not report the last input char as the unexpected
one). Also ended up moving diagSingleton out of the header file since
nobody outside of dfa.cpp should be constructing a DiagSet anyway.