Moving vector<string> formatter to test_util.h
Previously it was in dfa_test.h, but it will now be used by other tests.
Things in test_util.h is now moving to namespace oalex, while it was in
the global namespace in the past. This results in just one extra using
declaration.

We also removed redundant debug() calls from dfa_test.cpp, inlining the
function away into operator<<(ostream&,...).