Lexer::Fatal added
Causes UserErrorEx to be thrown, which is never caught by Lexer. It
isn't a new Diag::severity, since it is never actually stored in
Lexer::diags for *later* reporting. It's displayed to the user
immediately, likely before the entire program exits.

It is meant for errors such as maxLineLength violations, where no
sensible recovery is possible without running into resource constraints.