Added lexBracketGroup
Taking inspiration from YAML, we will not have any indentation-sensitive
component inside brackets. So it is possible for lexBracketGroup() to
mostly be a lexer, tokenizing and skipping space and comments. In a
bracket, we can't have section headers, but we can have quoted strings,
numbers, and identifiers.

This includes a completely new set of parallel utilities for writing
tests: lexer_matcher.{h,cpp}. It makes it easy to write tests where we
have to indicate what parse tree we are expecting.