Change lookahead utilities to lexing versions
lookaheadWord and lookaheadOperator is replaced with lexWord and
lexOperator. Almost identical, but modifies `i` parameter. This will
allow us to reuse this for lexing bracketGroup.

We also made the semantic change of not calling brackets and quotes
"operators" for lexing purposes. Their handling is sufficiently
different that it doesn't make sense to keep them grouped together.
This is why lookahead() now explicitly calls isbracket() and isquote().
Again, this will help processing for lexing bracketGroup.