Don't forget newline positions
class Input tries to optimize memory usage a bit too much, since we need
its list of newlines_ throughout the compilation process. Essentially,
we have one of two choices:

  * Forget newlines here, but keep row/col/byte_offset for every token
    type in lexer.h.
  * Keep newlines here, let token classes keep only byte_offset. We can
    convert to row/col using Input::rowCol() whenever we need.

Even if we were worried about memory usage (we are not), the latter is
better. There are usually multiple tokens in a line, and they will be
gaining two extra size_t each. Compare that to a single size_t per line.

The real bonus is lots of complex code being deleted.