Move class Lexer to InputDiags
The class Lexer doesn't actually have anything to do with lexing. It's
just the Input object and associated diags, both of which have moved to
runtime/ folder by now. It was therefore renamed to InputDiags and moved
there as well.

Revealed a good number of missing using declarations in
{lexer,jsonloc}_test.cpp. Turns out they were hidden by
argument-dependent lookups.