Adding automatic getch() conversion for Skipper.
This was a TODO in skipper_test.cpp. We may move it to the library later
if this becomes useful. For now, it only serves the tests.

I'm also slightly surprised that I didn't catch the 7-->8 change
earlier. Did I not run unit tests when I added Haskell-style comments?
That's odd.