Style fix: consistently class-ify npos
It's all the same value anyway, so this probably doesn't matter. But it
should be consistent. At some point I had decided I should use obj.npos,
but wasn't consistent in that either. I feel like it'd be easier for me
to stick to using appropriate class names, to indicate what type of
object it is indexing into.

I wish indices were type-safe, and not all ints.