Change plans for handling blank lines.
Previous plan was to stop at the first newline preceding a blank line,
and then provide a Skipper::blankLines() utility for the last of blank
lines. Not only did this interact poorly with last comments ending
mid-line, it also encouraged unbounded lookahead. The gap between
indices returned by acrossLines and blankLines is unbounded, and we
would have to remember it inside Input::buf_.

New plan: return the newline preceding the next non-blank line, as long
as we have seen at least one blank line. It is enough to indicate the
presence of a Markdown-style paragraph break as long as clients are not
too picky about which newline character they are receiving.