Generalize fuzzer equality check
Soon, it will become much harder to do a direct string comparison on a
parse-print test. Once we start accepting inputs like "[ab]([cd][ef])"
with redundant parenthesis, we will start optimizing them away during
the very first parse. Unfortunately, this does mean we will miss any
mistakes made on the first parse.

So this commit keeps old tests as is, so we can keep the stricter test
on simpler cases. But for the more complex cases, we will settle for AST
comparison.