Implement escape codes in string
Refactored parseEscapeCode() to take in a list of custom metacharacters,
so it can be used both in the context of character classes and strings.

Fixed a few quirks with escaping during pretty-printing as well. E.g.
closing brackets and braces don't really have to be escaped, since they
can't be nested. Added a few metacharacters like forward slash and
parenthesis. needsEscaping() now supports '\n' and '\t', to go with
the parsing changes above.

Unfuzzed for now.