Limit repetition depth
While we do have kMaxDepth limiting depth of parenthesis groups in
regex, we still accept a string of redundant repetition operators.
Because of the postfix nature of these operators, they don't cause
a stack overflow during parse, since they simply modify the latest
on-heap structure in-place.

However, this still caused stack overflows in later *use* of this
regex.

This was found by fuzzing.

Removed old TODO.