Refactor out interface InputPiece
struct Diag shouldn't need to depend on class Input. Refactoring it out
will allow us to add new Diags even after the lexical phase is over and
the Input is 'forgotten'. All we need is some way to convert from byte
positions in a piece of string to the `rowCol` position.