Making Resetter less error-prone
Previous approach was error-prone. We fix this by distinguishing the
offset we are reseting to vs. the offset returned by rst.start().

Here is an example where error was easy.

  auto lexSomething(InputDiags& ctx, size_t& i) {
    Resetter rst(ctx, i);
    // Do something and then increment i.
    ++i;
    rst.markUsed(i);
    // Do something else
    rst.markUsed(i);   // Easy to forget.
    return something;
  }

Forgeting the second markUsed() would have caused a "leak", in that
nothing would get forgotten from that point on. E.g. We could mark
index 0 to 9, forget to markUsed 10 to 15 later, and then any markUsed()
call from 16 onwards in a later function would be ignored. This is
undesirable.

Now, we ensure that we always reset i to the last index used.