Have QuotedString derive from InputPiece
This finally allows us to locate parts of a QuotedString in source for
error messages. This transforms QuotedString into a class rather than a
struct. Members previously public become private, which forces us to
significantly rearchictect the new pieces of lexer_test.cpp.

The one semantic change is that we no longer support the lack of
trailing newlines in lexIndentedSource(). Previously, we were appending
an extra '\n' in such cases. This makes location-tracking harder when we
are introducing new characters that weren't in the input to begin with!