Explain QuotedString representation reason
It's a fairly complicated class that is supposed to be used as "just a
string". Explain how the complexity is introduced by the need to carry
around location information.

Chances are very high I will waste time trying to optimize this in
future, so it's worth trying to preempt those attempts.