Make DiagDest members public
The methods here aren't really enforcing any useful invariant anyway.

We need to be able to add Errors to const QuotedString, which is not
currently possible since Error() is a non-const method. Eventually, the
Error/Warn/Info will become free templated functions that won't be tied
to the const-ness of QuotedString.