Reuse ctx among test QuotedStrings
This commit replaces the quote() macro with a new fquote() pattern.

Previously, quote() would create a temporary InputDiags to parse it and
extract a new QuotedString object for use in a test. That temporary
InputDiags would then go out of scope, and be destroyed. This had some
unintended consequences:

  * QuotedString kept a dangling pointer to InputDiags::diags. This
    was not a problem when quote() was first added, since QuotedString
    had no such pointer member. This was later added to facilitate
    error-reporting.
  * On the topic of error-reporting, every quote would be coming from
    "line 1 col 1" in every related error message, since they were all
    plucked from fresh strings passed to quote(), and not from
    substrings of a larger file.
  * Because the related diags vector was no longer available after
    quote() returned, it was also impossible to invoke
    assertEmptyDiags() after tests.

To remedy this, the test now starts with a large string with an
associated InputDiags ctx object that stays alive throughout the test.
An associated fquote() lambda is also produced that finds relevant
substrings from the larger input and constructs QuotedInput. This will
also allow us to invoke assertEmptyDiags() when we add diagnostics in
the following commits.

Alternatively, we could have also designed fquote() to accept an integer
i instead of a substring, so as to find the i-th string literal. But the
substring option was considered more readable so that the reader didn't
have to keep referring back to the full text input and count through it.
It was also considered less error-prone in the face of insertion and
deletion of quotes in the larger test input.

Yes, this required us to repeat test substrings in two places, but that
was considered a reasonable tradeoff for readability.