Produce error messages in matchAllParts()
This commit replaces every `return nullopt` in matchAllParts() with
appropriate error messages. This also allows us to use
assertEmptyDiags() to check that we are *not* producing error messages
on success.

This also allowed us to add recovery on an unterminated segment,
resolving an old TODO.

Tests for these new error messages also revealed a bug in
matchAllParts(spatt, s), where we had reversed the order of a
comparison.