Store ctx pointer in QuotedString
It is sometimes more convenient to pass around rowCol() information as a
single index instead of a pair. This can be done by translating all
QuotedString-internal offsets into ones global in a given ctx.input.

In the next few commits, we will replace QuotedString::row_col_map_ into
QuotedString::index_map_. Having a reference to the entire ctx object
will allow us to produce rowCol() on-demand, instead of storing it
internally.

This is the first change that only changes the diags pointer in
QuotedString to point to the parent ctx object. The later ones will
change row_col_map_ representation.