s/RowColRelation/IndexRelation/g
It no longer makes sense to call this structure RowColRelation when it
no longer stores relationships involving row and column.

We also moved the quotePos to come first, since that matches previous
uses of makeRowColRelation(), a function that can now be deleted.
Instead of makeRowColRelation(), it is simple enough to directly use
IndexRelation{} brace initialization.

Made a few other renames for consistency's sake, e.g.

  row_col_map_ --> index_map_
  rcmap --> imap