Add QuotedString::subqstr()
We need to be able to extract substrings from QuotedString in
labelParts(), since it's job is to split strings.

This is intentionally named to be different from QuotedString::substr().
subqstr() will be a much more heavyweight operation, and should be used
only if needed. This function has to recalculate its index_map_ since it
now starts in a different position of user input. It will reuse some of
the same .inputPos values, but now map them to a different .quotePos.