Add InputPiece::bol()
Previously, bol() was a member of just `class Input`. This commit adds
the same functionality to QuotedString as well, so it can be exposed in
the parent interface InputPiece.

This is the more complicated of a few utility methods that need to be
moved up from Input to InputPiece, so that we can have class Skipper
operate on both subtypes of InputPiece. Today, that class only works on
Input.

As part of this change, we ended up refactoring the std::upper_bound()
calls in lexer.cpp implementation.