Port Skipper methods to use InputDiagsRef
This resolves an old TODO in skipper.h, and will allow us to use Skipper
in template.cpp. One minor change is that we no longer call
input.forgetBefore() in Skipper::acrossLines(). We can revisit this
issue if it becomes a problem. I don't expect any problem in the short
run.