Skipper to indicate unfinished comments with npos
Previously, it was producing a diagnostic error message, and was pretty
loose about whether the return value was npos or just "any value greater
than size()". Fortunately, this class isn't used widely enough for a
semantic change to be a problem.

This change will allow us to customize the diagnostic produced on the
caller-side. E.g. We can still produce a "Comment never ends" message if
the input spanned the whole file. On the other hand, it could be
something weirder like "Placeholders not allowed in comments" if the
input is part of a template.