Implement parsing '|' branches
The primary change is that we now invoke close_curbranch() on every '|'
token, in addition to the one in ']'. This moves the current set of
tokens into prevbranches(), and clears curbranch() for everything after
the '|'. close_curbranch() now has extra logic for choosing an
appropriate error message based on the stack context.

If close_curbranch() does see an error (on empty branches), we stop
further parsing.