Add support for repeats and folds
Preliminary support for ellipsis (...) handling. This is the
implementation approach: when finishing up a branch of concatenated
tokens, we pass it through an additional ellipsis processing phase
implemented using the new internal function repeatFoldOnEllipsis().
Until a branch is finished, an ellipsis is added to the token stream
like any other non-special token.