Define new heuristic for ellipsis expansion
This commit does not yet change the behavior of repeatFoldOnEllipsis(),
which will happen over the next few commits. Instead, it introduces a
new function rolloutEllipsis() that implements the new heuristic.

This rolloutEllipsis() is a function template with two intended
instantiations: one that operates on a sequence of characters and the
other that operates on a sequence of AST nodes. The former is used here
for testing as part of rolloutEllipsisForTest(). The latter will be
introduced into repeatFoldOnEllipsis() as part of the next few commits.

The algorithm tries to find all possible interpretations of an ellipsis,
and is quick to jump on the gun and scream "Ambiguous". Only if a
somewhat unique instantiation exist do we parse it. Ambiguity can be
almost always avoided by adding a few more periods of examples to the
input.