Use newer heuristic in repeatFoldOnEllipsis()
Instead of calling getSurrounding() twice, it now calls
rolloutEllipsis() once. This allows us to resolve the old feature wish
that was marked as a TODO. As the new test cases show, we can now handle
more cases like, for instance "a a ... a a" which was previously
ambiguous. Other test cases were already added as part of
testRolloutEllipsis().