Make Regex variants all const
This was mainly an exercise to see where I am missing const but should
have had it from the start. So the actual type definition in regex.h
might be reverted one day. The only place where this constness was
slightly awkward was in regex_io.cpp, where we need to manipulate these
objects until we return them.

Another approach we may take later is to leave regex.h const as is, but
make a separate MutableRegex type internal to regex_io.cpp. For now,
that wasn't necessary.