Implement RegexAnchor::wordEdge codegen
Word edge anchors requires the use of RegexOptions. They are declared as
a separate defaultRegexOpts() global to all rules in the RuleSet. For
now, we are assuming the generation of only a single RuleSet at a time.

We are also resolving an obsolete TODO about "global using"
declarations. It seems better not to pollute the global namespace for
now.