Use placeholder names for result variables
ConcatRule codegen was using the wrong names for local variables in
generated code. This became a problem if the same rule is used twice in
the same name (e.g. parsing two identifiers as part of an assignment).
Instead of storing them as resIdentifier, we must use resLhs and resRhs
to avoid a name collision. We need to use placeholder names, not rule
names.

The frontend should always guarantee unique placeholder names in a given
rule.

This bug was discovered as I was trying to call the same Skipper
function multiple times.