Hide second stage of template processing
We want to disallow run-on templates like "operandbinopoperand", at
least for templates that ignore whitespace. This requires us to perform
additional processing between marking labels and word-splitting, which
should not be exposed to the library user.

This is the first commit in that vein, and makes no functional change
yet. It renames the old tokenizeTemplate() to
tokenizeLabelledTemplate(), hides it in the .cpp file by marking it
static, and then exposes a new tokenizeTemplate() function in the header
that combines labelParts() and the newly renamed
tokenizeLabelledTemplate().

The next commit will add word-boundary checks.