labelParts() can avoid word-splitting
It can now accept an additional parameter about what is or is not a
"word char". In order to accommodate the new unit tests, we had to make
the findIdent() helper function a bit more flexible, so that it no
longer latches onto the first occurrance of a substring, but keeps
scanning.