Use RuleSet for parsing
Previously, we were manually calling hasPrefix() on the input, and
ignoring any RuleSet parameter. Now, we actually populate the RuleSet
parameter returned from parseOalexFile(), and call eval() on it.

Change the warning in regex_io.h to indicate it is safe outside of tests
as well, as long as it's not used on user-input. Duplicated the warning
in oalex_main.cpp as well, and provided alternate suggestion in
regex_io.h.