Improve lexNextLine() error handling
The previous implementation didn't anticipate lexQuotedString()
attempting recovery and changing i even on failure. This one does, by
committing to a particular type of token based on the first character.

lexNextLine() now correctly returns an error if any of the quotes fail
to parse. It does not change i when it does.