Rewrite parseOalexSource to allow string rules
The capability of the parser has now expanded slightly. Instead of
allowing just "require_politeness", it now also supports single-string
rules, of the form `a := "hello"`.

So the input file can now have multiple lines of rules, but only the
first one actually gets used. These rules can either be this
single-string rule, or the original `require_politeness` on a line all
by itself. The former allows the parsed string to be customized, for
once.