Record how a GluedString object was created
Different client code may treat them differently based on, e.g. if it
was a single-quoted string or double-quoted string. The 'subqstr' is
just a tag that indicates we forgot the original type. There is no
reason to propagate this information for now.

The "single-quoted" enum value squoted hasn't been added yet. It will be
there in the next commit.