Add tests to `require_politeness`
This adds in a test so that Expectation::matches() can be finished. But
this is still one hardcoded test, and not something specified in oalex
input source. As a result, `oalex test` still needs to be tested
manually.

My next job will be to start parsing extremely simple test cases.