Use inheritance instead of variant in Regex
C++ sum types are not nearly as efficient as other languages, and effect
they have on compilation time is ungodly. This commit changes all of
that for a more conventional inheritance-tree structure for ASTs. This
Regex type, however, is not the only use of std::variant in the
codebase. As a result, it only shaves off a few seconds from the total
build type. It is, nevertheless, a useful step towards completely
eradicating std::variant from here. We can reconsider this in 15 years'
time if implementations improve.

That said, this is probably the worst-organized, largest commit I have
allowed myself to author in a long time. I would have done it in a more
piecemeal fashion under better circumstances. Most changes, however, are
mechanical:

  Regex, optional<Regex>
    -->  unique_ptr<Regex>
  if-else ladder of get_if<T>(&regex)
    -->  switch-case over regex.nodeType
  RegexAnchor
    --> RegexAnchor::AnchorType

Similar changes to JsonLoc, Template, and other variants will have to be
done in a more organized manner. E.g.:

  1. Define the new types.
  2. Define conversions from old to new for transition crossovers.
  3. Convert one function at a time.