Support json test outputs
The only rule that produces such outputs now is `require_politeness
jsonized`, a dummy rule. Most of the changes in this commit actually
went into parsing this dummy command, inside parsePolitenessDirective().

The useful change is actually the small diff inside parseExample(),
where we can push in a new type of Expectation along with the parsed
example. Expectation::SuccessWithJson.

The part of matching JsonLoc objects was mostly done in prior commits
with the introduction of operator==. Here, we just invoke that in
Expectation::matches().

testdata/17-bad.oalex now correctly complains about the output, instead
of screaming "unimplemented".