Add std::monostate as a Rule variant
This will allow delayed rule construction, e.g. for reserving a named
rule slot when parsing a RuleSet BNF rule.

Context: the BNF-style rules `lhs := rhs` were introduced to quickly
iterate through new rule types without spending a lot of time parsing
them. This "quickly" means we don't really want a heavyweight set of
data structures for the BNF rules separate from those in codegen.h.

However, they were not as lightweight as we would have liked it. codegen
data refers to each other by integer indices, not name.  We do want the
BNF rules to use names. Indices become cumbersome, if only because they
get messed up by anonymous rules.

So here we are. We are introducing a mechanism for reserving a rule slot
with a stable index as soon as we see them, perhaps on the rhs of some
rule. Later during the parse, we can fill it up when the name shows up
on the lhs of some other rule.

Using names will also allow us to seamlessly refer to and from BNF rules
and other rule types that we will later introduce.