Autogenerate rule names in frontend when needed
The code generator requires most rules to have names, so that we can
generate function names in output c++ file. This change adds simple
names like "rule1", "rule2", whenever rules need names.

The only rules that don't need names are the ones that are implicitly
defined as a component of some other rule, such as a string literal
component in `Concat[]`.

This lack of names has not been an issue so far, since we are only
testing with eval(). However, we need to fix this before we add
codegen() (aka `oalex build`) in our automated tests.