Call templatize(), but don't use its value yet
Just the work of calling templatize() requires a fair bit of set up, in
terms of linking new files, new using declarations, new parameter
plumbing, etc. So it made sense keep this in a separate commit, while
the next one actually makes use of the result.

This commit checks the templatize() result with has_value(), but doesn't
otherwise use it yet.