Update header comments
We are not removing ConcatRule for now, since flattening code makes a
lot of assumptions we don't have with ConcatRule. E.g. the fact that all
JsonLoc::Map components must be nameless. This makes sense in the
context of flattening, but parseConcatRule() always produces names of
the form arg1, arg2, ..., argN.

Maybe we can remove ConcatRule once we are sure that we no longer need
it for quickly testing with Concat[] BNF rules.