Revise plans regarding OrRule return value
The fact that OrRule can sometimes return a Map suitable for OutputTmpl
and sometimes cannot is becoming complicated. We cannot safely compose
it with OutputTmpl at compile-time. Instead, the plan now is to support
all manner of return types from OrRule branches, but wrap them up if
it's not a Map. This is exactly what we do for ConcatFlatRule children,
and is an accurate representation of how humans write parser code.