Package rules and firstUseLocs into a class
This should have been done a long time ago. I will likely send a few
more commits to move a few functions to be methods of the class, just so
we can maintain some simple invariants (like sizes being equal). But
this commit is just the initial packaging of fields.