rl.rules.size() --> rl.ssize()
This is another step in making rl.rules private. It adds a ssize()
method. As an added bonus, it removes a lot of sign-casting.

Some instances of rules.size() remains in member methods of
RulesWithLocs. Changing them to use this->ssize() would have caused a
sign mismatch. All the better. Those lines can be completely removed
once rules is a private member.