Make RulesWithLocs::rules private
This finishes the current refactoring bout. Once the frontend is done
with making all the rules, it invokes a new method releaseRules() to
export the rules out. The class still maintains size equality by
clearing firstUseLocs_, effectively reinitializing the object.

This has the added benefit that I can now be less paranoid about a
size-mismatch from creeping in.