findOrAppendIdent() now accepts an Ident
Previously, we had to provide separate string and location parameters.
The locations weren't even consistently filled in everywhere. We now
accept Ident objects, which allows us to contain the use of
preserveCase() a bit more.

This change led me to discover a few other issues, which have now been
marked with TODOs and FIXMEs in comments.