Restrict Rule renaming
The only remaining use of this is for giving generic rule1, rule2, etc.
names for anonymous Rules at the end. Even that should become
unnecessary once we have better name generation.

The setter Rule::name(string_view) is also renamed to
Rule::deferred_name(string_view) to indicate that it's not an ordinary
setter.