Don't require naming a rule twice
Previously, we were initializing each rule in parseBnfRule twice,
without realizing it. Once was for defineIdent(), and the other was when
overriding it with rule specifics.

We now have a new deferred_assign() that lets us override just the
specifics. This led to some simplification through a few functions that
no longer need the rule name (or "ident") to be carried around. Indeed,
we were passing in an empty rule name in many cases anyway!

While we are at it, this commit also deletes the Rule assignment
operators to make sure this isn't still being used anywhere. A test case
in codegen_gen_test.cpp was simplified as a result.