Impl Ident::parse(WholeSegment)
This will help us compose parsing Ident with lexBracketGroup(). When we
parse ident, it turns out that sometimes we have already tokenized the
stream some other way.

This commit factors out most of the body of Ident::parse(ctx, i) into
a new Ident::parseFromString(). This allows us to reuse most of it in
the new Ident::parse(WholeSegment) being introduced.