Stop using InputDiagsRef::input in ident.cpp
We want to turn InputDiagsRef into a "DiagsDest" that is only used for
producing diags. In the process, this input field will become a private
member. This removes the only public usage of this field outside of
diags.h.

Changing it to InputDiags& makes sense since, going forward, this will
be the only type used by parsers directly processing bytes from the
input string. In the past we used InputDiagsRef when we were parsing
from bits of GluedString objects, but we are not doing that anymore.

The other Ident constructor can keep using InputDiagsRef since it
doesn't use the input member. It will later be renamed to DiagsDest to
reflect its true use.