Add parseGenerated()
This method is meant to be used for generated identifiers, either as
part of the main binary or in tests. It dies on invalid input, since it
is expected to have sanitized inputs only.

The reason for introducing it now is so that we can start migrating
Rule::name_ from being of type std::string to oalex::Ident. Transition
period will make heavy use of the fact that for all string s,

  parseGenerate(s).preserveCase() == s