Change return type of Rule::name() to Ident
This completes the migration of Rule::name_ from being of type
std::string to oalex::Ident.

We convert back to string with preserveCase() just before outputting the
name, but keep it as Ident otherwise. This eliminated a couple more
redundant Ident-to-string conversions in findOrAppendIdent() and
defineIdent().

One interesting place where we are still relying on strings is in
Examples. Added a TODO for that.