Don't flatten all OrRules
This change adds OrRule::flattenOnDemand, which dictates whether
ConcatFlatRule will have its flattening effect on it. It will allow us
to write OrRules that do not worry about whether they must return a
JsonLoc::Map or not. This is accomplished by using this new field as the
return value of makesFlattenableMap() for OrRules.

Previously, it was assumed that they must always return maps. This was
because we were only thinking of using these rules as part of patterns,
not directly defined as a multi-match rule. We will now use it as the
workhorse for multi-match rules.

This commit also simplifies the previously verbose comment block in the
OrRule definition.

New tests now exercise makesFlattenableMap() for both eval() and
codegen(). This resolves the FIXME introduced in the previous commit.