Use consistent flattening in codegen and eval
Previously, eval() was deciding to flatten based on whether child
returns a map. That is not consistent with what codegen() is doing. This
change uses makesFlattenableMap() in both places.

For now, it doesn't make a difference as long as all OrRule components
of ConcatFlatRule return maps like they are supposed to. But in the next
commit, we will have variants of OrRule that do not return a map
consistently. That's when we will see an observable difference in
behavior.

No changes in test, since we have no observable changes in behavior (for
valid inputs).