Test that lexNextLine() ends at bol()
lexNextLine() requires parsing to start at beginning of a line.
Therefore, consecutive calls to lexNextLine() can only succeed if each
one ends at the bol() of the next line. I just spent half an hour
examining if that is indeed so.

Adding a test to ensure it stays that way.