Rule::needsName() accepts lookahead target status
Whether or not a rule needs a name now depends on not just the support
for anonymous rules we have in codegenParserCall(), but also in
codegenLookahead().

Right now this bit is unused because the two functions happen to support
the same set of rules, so we don't need to consult it. But that was not
true just a few commits ago, and is unlikely to remain true forever.

The reason this has not caused a problem so far is that the frontend
currently disallows unnamed lookaheads. This won't remain true going
forward. As far as codegen goes, the previous needsName() behavior was a
bug that wasn't triggered by the caller (the frontend). The bug went
away when we added special handling for string and WordPreserving. This
commit makes it less likely for the bug to be reintroduced.