Stop checking hasError() in lookaheads
We were previously failing peekMatch() on hasError() in the hope that
a "better" branch might later be a better match. This turned out to be
too confusing for the oalex user.

If a rule deems something to be a success, other rules shouldn't look at
side-channels like error diagnostics to second-guess that. We will need
error-producing successful matches to perform error-recovery properly.