Create Ident-wrapper eagerly
We currently need Ident-parsing results to be map-typed to interact
properly with any enclosing ConcatFlatRule. This was previously done
lazily, only when an enclosing ConcatFlatRule was being emitted. This
commit changes it to happen eagerly, whenever we are compiling Ident
patterns.

This helps because the previous solution is now being used in more
places. appendPatternConcat() and appendPatternOrList() both tries to do
this by checking if a child is "expected" to return a map. This
expectation-detection gets finicky, as it is yet another place that
needs to stay in sync with Rule result types. Moreover, this usage will
have to proliferate get ready to compile PatternOptional and other
variants.

Instead, it's better to always wrap Ident results in maps pre-emptively.
This allows us to delete the patternName() function, reducing the need
to stay in sync with Rule result types.