Separate out glue terms in fold rule tests
Soon, I will be introducing a dedicated field for part and glue, so we
no longer need a full-featured ConcatFlatRule to be embedded inside a
LoopRule. This change of unit tests is meant to help that transition.

I've also added descriptive messages to assertEqual() wherever I found
it useful for testing.

No change in any code outside of tests.