Introduce glueidx for eval()
Previously, the last component of children.comps could optionally be a
glue component. This used to be indicated by having breakBefore set to
a value less than ssize(children.comps). This commit separates it out
into its own component indicated by a new field, glueidx. This is only
done for eval() now, and the codegen() change will be done in a later
commit. For now, we will constrain the breakBefore field to always be
ssize(children.comps) redundantly, indicating none of those are glue
components.

Given the constraint, the main loop will no longer need to handle the
case where childi == loop.breakBefore, since that's impossible. The
presence or absence of a glue component is determined by glueidx == -1.
Finally, if the glue component is present, it is handled separately
outside the main loop.