Reformatting
It turns out that almost all functions in this file are static. It makes
more sense to enclose the whole thing in namespace {} than to keep
marking everything as static.

I also took this opportunity to put all the return types in a separate
line, like I had always wanted.