Delete soleIdent()
Given the conventions I was already following for pattern-compilation,
it turned out that the convoluted logic of this function was entirely
unnecessary. They were useful back when Pattern-compilation would use
ConcatRule, but became redundant with our move to ConcatFlatRule. This
will in turn also simplify how we compile PatternRepeat and PatternFold
objects.

I also added a large block of comment so that I am less confused the
next time I am adding a new kind of Rule or Pattern.