s/makesFlattenableMap/resultFlattenableOrError/g
This allows us to fold in ErrorRule into this as well. This is just
recognizing the fact that all rules can potentially return
JsonLoc::ErrorValue, and that does not disqualify them from this
function. Again, this is meant to make OrRule stricter about what kind
of child components it accepts.