Delete LexSegment
It is not really serving any useful purpose. It has not stopped any
notable mistakes, and now it seems we may actually *want* to define some
of these types in a different header file. A lot of other headers use
GluedString and WholeSegment, but has no need to include so many other
headers transitively.