Move WholeSegment to segment.h
Previously, we had a lot of transitive dependencies being pulled into
ident.cpp and ident_test.cpp. Most of the diff is for pulling out
WholeSegment of namespace oalex::lex to namespace oalex.

Apparently, the Error() helper function was using argument-dependent
lookup. So this change required it to be pulled out of namespace lex as
well. We took this opportunity to also detemplatify it.