Remove GetFromString() and Unixify() when possible
It's been a long time since we have had a string constructor for Input.
So these helpers haven't been necessary for a while. The only remaining
exception is assertProducesDiag(), which is used in testLineTooLong().
We expect Input::operator[] to raise an error in those cases, but that
only happens if a callback is used for inputs.