Move function bodies out of diags.h to diags.cpp
These functions are never in the hot path, and therefore do not need to
be inlined. For now, they are needlessly adding to the compilation time.

This also removed util.h from being transitively included here. For now,
this include had to be re-added to a few unrelated files (e.g.
ident.cpp, codegen_test_util.h, and parser_helpers.h).

Result: we shaved off a good 20s off the total build-time of 1m30s.