Replace JsonTmpl::substitutionsOk()
When this method was copied from JsonLoc, it was used after
substitutions to see if any more were still necessary. Now, however,
JsonTmpl represents the pre-substitution state.

Now that the two types are different, it also no longer makes sense to
have checks for stPos and enPos that are based on user input, while the
positions in this class represents oalex source locations.