Pre-resolve Ident lookups in pattern compilation
Previously, the PatternToRulesCompiler would directly look up the list
of rules whenever it encountered a placeholder Ident. Now, it instead
consults a map that is precomputed with all the rule indices.

This will allow the caller to customize the lookup resolution, which
will be useful when add local identifiers that will not belong in the
list of global rules.