Prepare parseRule() for where clause
Right now, this function has a rigid grammar that expects `outputs`
right after the pattern. Soon, it can also be a `where` stanza. To
prepare for that, we are switching from a straight-line code to a while
loop, just so we can tolerate them in any order.

Introducing a new `sawOutputsKw` to make sure we don't see `outputs`
multiple times. While this does not introduce any new feature, we do
still introduce a new error message.