Separate DefinitionInProgress from UnassignedRule
This change allows us to raise an error if defineIdent() is called
multiple times on the same Ident. This resolves a TODO in
compiler_test.cpp. We now have two different notions:

  * DefinitionInProgress, when we are still expecting a
    deferred_assign() very soon. We are merely trying to figure out what
    to define, rather than whether to define a new rule.

  * UnassignedRule for when a rule was already used somewhere, but we
    haven't seen it defined yet. It may be defined later in this file.

These two concepts were both called UnassignedRule previously. A
consequence of this change is that we should no longer have anonymous
UnassignedRule objects.